+++ src/main/java/nl/enjarai/doabarrelroll/net/HandshakeServer.java
@@ -1,9 +1,11 @@
 package nl.enjarai.doabarrelroll.net;
 
+import com.mojang.serialization.Codec;
 import com.mojang.serialization.JsonOps;
 import io.netty.buffer.Unpooled;
 import net.minecraft.network.PacketByteBuf;
 import net.minecraft.server.MinecraftServer;
+import net.minecraft.server.network.ServerPlayNetworkHandler;
 import net.minecraft.server.network.ServerPlayerEntity;
 import nl.enjarai.doabarrelroll.DoABarrelRoll;
 import nl.enjarai.doabarrelroll.util.DelayedRunnable;
@@ -11,15 +13,21 @@ import nl.enjarai.doabarrelroll.util.DelayedRunnable;
 import java.util.ArrayList;
 import java.util.Map;
 import java.util.WeakHashMap;
-import java.util.function.Supplier;
+import java.util.function.Function;
 
-public class HandshakeServer<T extends SyncableConfig<T>> {
-    private final Supplier<T> configSupplier;
-    private final Map<ServerPlayerEntity, HandshakeState> syncStates = new WeakHashMap<>();
+public class HandshakeServer<T extends SyncableConfig<T> & ValidatableConfig> {
+    private final ServerConfigHolder<T> configHolder;
+    private final Map<ServerPlayNetworkHandler, HandshakeState> syncStates = new WeakHashMap<>();
     private final ArrayList<DelayedRunnable> scheduledTasks = new ArrayList<>();
+    private final Function<ServerPlayNetworkHandler, Boolean> getsLimitedCheck;
+    private final Codec<T> transferCodec;
+    private final Codec<? super T> limitedTransferCodec;
 
-    public HandshakeServer(Supplier<T> configSupplier) {
-        this.configSupplier = configSupplier;
+    public HandshakeServer(ServerConfigHolder<T> configHolder, Function<ServerPlayNetworkHandler, Boolean> getsLimitedCheck, Codec<T> transferCodec, Codec<? super T> limitedTransferCodec) {
+        this.configHolder = configHolder;
+        this.getsLimitedCheck = getsLimitedCheck;
+        this.transferCodec = transferCodec;
+        this.limitedTransferCodec = limitedTransferCodec;
     }
 
     public void tick(MinecraftServer server) {
@@ -28,17 +36,24 @@ public class HandshakeServer<T extends SyncableConfig<T>> {
     }
 
     public HandshakeState getHandshakeState(ServerPlayerEntity player) {
-        return syncStates.getOrDefault(player, HandshakeState.NOT_SENT);
+        return getHandshakeState(player.networkHandler);
     }
 
-    public PacketByteBuf getConfigSyncBuf(ServerPlayerEntity player) {
+    public HandshakeState getHandshakeState(ServerPlayNetworkHandler handler) {
+        return syncStates.getOrDefault(handler, HandshakeState.NOT_SENT);
+    }
+
+    public PacketByteBuf getConfigSyncBuf(ServerPlayNetworkHandler handler) {
         var buf = new PacketByteBuf(Unpooled.buffer());
 
         // Protocol version
-        buf.writeInt(1);
+        buf.writeInt(2);
 
-        var config = configSupplier.get();
-        var data = config.getTransferCodec().encodeStart(JsonOps.INSTANCE, config);
+        // Config data
+        var config = configHolder.instance;
+        var isLimited = getsLimitedCheck.apply(handler);
+        Codec<? super T> codec = isLimited ? limitedTransferCodec : transferCodec;
+        var data = codec.encodeStart(JsonOps.INSTANCE, config);
         try {
             buf.writeString(data.getOrThrow(false, DoABarrelRoll.LOGGER::error).toString());
         } catch (RuntimeException e) {
@@ -46,54 +61,58 @@ public class HandshakeServer<T extends SyncableConfig<T>> {
             buf.writeString("{}");
         }
 
+        // Limited status
+        buf.writeBoolean(isLimited);
+
         return buf;
     }
 
-    public void configSentToClient(ServerPlayerEntity player) {
-        syncStates.put(player, HandshakeState.SENT);
+    public void configSentToClient(ServerPlayNetworkHandler handler) {
+        syncStates.put(handler, HandshakeState.SENT);
 
-        var config = configSupplier.get();
+        var config = configHolder.instance;
         if (config.getSyncTimeout() != null) {
             scheduledTasks.add(new DelayedRunnable(config.getSyncTimeout(), () -> {
-                if (syncStates.getOrDefault(player, HandshakeState.NOT_SENT) != HandshakeState.ACCEPTED) {
+                if (syncStates.getOrDefault(handler, HandshakeState.NOT_SENT) != HandshakeState.ACCEPTED) {
                     DoABarrelRoll.LOGGER.warn(
                             "{} did not accept config syncing, config indicates we kick them.",
-                            player.getName().getString()
+                            handler.getPlayer().getName().getString()
                     );
-                    player.networkHandler.disconnect(config.getSyncTimeoutMessage());
+                    handler.disconnect(config.getSyncTimeoutMessage());
                 }
             }));
         }
     }
 
-    public HandshakeState clientReplied(ServerPlayerEntity player, PacketByteBuf buf) {
-        var state = getHandshakeState(player);
+    public HandshakeState clientReplied(ServerPlayNetworkHandler handler, PacketByteBuf buf) {
+        var state = getHandshakeState(handler);
+        var player = handler.getPlayer();
 
         if (state == HandshakeState.SENT) {
-            var protocolVersion = buf.readInt();
-            if (protocolVersion != 1) {
-                syncStates.put(player, HandshakeState.FAILED);
-                DoABarrelRoll.LOGGER.warn(
-                        "Client of {} sent unknown protocol version, expected 1, got {}. Will attempt to proceed anyway.",
-                        player.getName().getString(),
-                        protocolVersion
-                );
-            }
-
             try {
+                var protocolVersion = buf.readInt();
+                if (protocolVersion < 1 || protocolVersion > 2) {
+                    syncStates.put(handler, HandshakeState.FAILED);
+                    DoABarrelRoll.LOGGER.warn(
+                            "Client of {} sent unknown protocol version, expected range 1-2, got {}. Will attempt to proceed anyway.",
+                            player.getName().getString(),
+                            protocolVersion
+                    );
+                }
+
                 if (buf.readBoolean()) {
-                    syncStates.put(player, HandshakeState.ACCEPTED);
+                    syncStates.put(handler, HandshakeState.ACCEPTED);
                     DoABarrelRoll.LOGGER.info("Client of {} accepted server config.", player.getName().getString());
                     return HandshakeState.ACCEPTED;
                 } else {
-                    syncStates.put(player, HandshakeState.FAILED);
+                    syncStates.put(handler, HandshakeState.FAILED);
                     DoABarrelRoll.LOGGER.warn(
                             "Client of {} failed to process server config, check client logs find what went wrong.",
                             player.getName().getString());
                     return HandshakeState.FAILED;
                 }
             } catch (IndexOutOfBoundsException e) {
-                syncStates.put(player, HandshakeState.FAILED);
+                syncStates.put(handler, HandshakeState.FAILED);
                 DoABarrelRoll.LOGGER.warn(
                         "Client of {} sent invalid config reply.",
                         player.getName().getString()
@@ -105,8 +124,8 @@ public class HandshakeServer<T extends SyncableConfig<T>> {
         return state;
     }
 
-    public void playerDisconnected(ServerPlayerEntity player) {
-        syncStates.remove(player);
+    public void playerDisconnected(ServerPlayNetworkHandler handler) {
+        syncStates.remove(handler);
     }
 
     public enum HandshakeState {