+++ src/main/java/nl/enjarai/doabarrelroll/util/MagicNumbers.java
@@ -1,0 +1,6 @@
+package nl.enjarai.doabarrelroll.util;
+
+public class MagicNumbers {
+    public static final double TORAD = Math.PI / 180;
+    public static final double TODEG = 1 / TORAD;
+}